+++ src/Plivo/Resource/Powerpack/PowerpackInterface.cs
@@ -241,7 +241,7 @@ namespace Plivo.Resource.Powerpack
 
 			return ExecuteWithExceptionUnwrap(() =>
 			{
-				var result = Task.Run(async () => await Client.Update<UpdateResponse<Powerpack>>(Uri +"Powerpack/"+uuid, data).ConfigureAwait(false)).Result;
+				var result = Task.Run(async () => await Client.Update<UpdateResponse<Powerpack>>(Uri +"Powerpack/"+uuid+"/", data).ConfigureAwait(false)).Result;
                 result.Object.StatusCode = result.StatusCode;
 				return result.Object; 
 			});
@@ -271,6 +271,7 @@ namespace Plivo.Resource.Powerpack
                 number_priority
             });
 			var result = await Client.Update<UpdateResponse<Powerpack>>(Uri +"Powerpack/"+uuid, data);
+
             result.Object.StatusCode = result.StatusCode;
 			return result.Object;
 		}
@@ -287,7 +288,7 @@ namespace Plivo.Resource.Powerpack
         /// <param uuid="uuid">UUID</param>
         /// <param name="offset">Offset.</param>
         public ListResponse<Numbers> List_Numbers(string uuid, string starts_with =null, string country_iso2 =null,
-        string type=null, uint? limit = null, uint? offset = null)
+        string type=null, uint? limit = null, uint? offset = null, string service = null)
         {
             var mandatoryParams = new List<string> {""};
             var data = CreateData(
@@ -298,7 +299,8 @@ namespace Plivo.Resource.Powerpack
                     country_iso2,
                     type,
                     limit,
-                    offset
+                    offset,
+                    service
                 });
 
 			return ExecuteWithExceptionUnwrap(() =>
@@ -325,7 +327,7 @@ namespace Plivo.Resource.Powerpack
         /// <param name="offset">Offset.</param>
          public async Task<ListResponse<Numbers>> List_NumbersAsync(
           string uuid, string starts_with =null, string country_iso2 =null,
-          string type=null, uint? limit = null, uint? offset = null)
+          string type=null, uint? limit = null, uint? offset = null, string service = null)
         {
             var mandatoryParams = new List<string> { "" };
             var data = CreateData(
@@ -336,7 +338,8 @@ namespace Plivo.Resource.Powerpack
                     country_iso2,
                     type,
                     limit,
-                    offset
+                    offset,
+                    service
                 });
             var resources = await ListResources<ListResponse<Numbers>>("NumberPool/"+uuid+"/Number", data);
             resources.Objects.ForEach(
@@ -362,7 +365,7 @@ namespace Plivo.Resource.Powerpack
         /// <param uuid="uuid">UUID</param>
         /// <param name="offset">Offset.</param>
         public uint  Count_Number(string uuid, string starts_with =null, string country_iso2 =null,
-        string type=null, uint? limit = null, uint? offset = null)
+        string type=null, uint? limit = null, uint? offset = null, string service = null)
         {
             var mandatoryParams = new List<string> {""};
             var data = CreateData(
@@ -373,7 +376,8 @@ namespace Plivo.Resource.Powerpack
                     country_iso2,
                     type,
                     limit,
-                    offset
+                    offset,
+                    service
                 });
 
 			// return ExecuteWithExceptionUnwrap(() =>
@@ -400,7 +404,7 @@ namespace Plivo.Resource.Powerpack
         /// <param name="offset">Offset.</param>
          public async Task<uint> Count_NumbersAsync(
           string uuid, string starts_with =null, string country_iso2 =null,
-          string type=null, uint? limit = null, uint? offset = null)
+          string type=null, uint? limit = null, uint? offset = null, string service = null)
         {
             var mandatoryParams = new List<string> { "" };
             var data = CreateData(
@@ -411,7 +415,8 @@ namespace Plivo.Resource.Powerpack
                     country_iso2,
                     type,
                     limit,
-                    offset
+                    offset,
+                    service
                 });
             var resources = await ListResources<ListResponse<Numbers>>("NumberPool/"+uuid+"/Number", data);
             resources.Objects.ForEach(
@@ -431,14 +436,15 @@ namespace Plivo.Resource.Powerpack
         /// <param name="uuid">UUID.</param>
         /// <param name="number">Number.</param>
 
-        public Numbers Add_Number( string uuid, string number, bool rent=false )
+        public Numbers Add_Number( string uuid, string number, string service = null, bool rent=false)
         { 
         var mandatoryParams = new List<string>{""};
            var data= CreateData(
             mandatoryParams,
             new
             {
-                rent
+                rent,
+                service
             }); 
         
     
@@ -458,14 +464,15 @@ namespace Plivo.Resource.Powerpack
         /// <param name="uuid">UUID.</param>
         /// <param name="number">Number.</param>
 		public async Task<Numbers> Add_NumberAsync(
-			 string uuid, string number, bool rent=false)
+			 string uuid, string number, string service = null, bool rent=false)
 		{
         var mandatoryParams = new List<string>{""};
                 var data = CreateData(
                 mandatoryParams,
                 new
                 {
-                    rent
+                    rent,
+                    service
                 });
            
 		var result = await Client.Update<Numbers>(Uri + "NumberPool/"+uuid+"/Number/"+number+"/", data);
@@ -614,19 +621,35 @@ namespace Plivo.Resource.Powerpack
         /// <returns>The Number resource.</returns>
         /// <param name="uuid">UUID.</param>
         /// <param name="number">Number.</param>
-        public Numbers Find_Number(string uuid,string number)
+        /// <param name="service">Service.</param>
+        public Numbers Find_Number(string uuid,string number, string service = null)
         { 
+            var mandatoryParams = new List<string>{""};
+            var data= CreateData(
+            mandatoryParams,
+            new
+            {
+                service
+            }); 
+
             return ExecuteWithExceptionUnwrap(() =>
 			{
-				var numberpoolResponse = Task.Run(async () => await GetResource<Numbers>("NumberPool/"+uuid+"/Number/"+number).ConfigureAwait(false)).Result;
+				var numberpoolResponse = Task.Run(async () => await GetResource<Numbers>("NumberPool/"+uuid+"/Number/"+number, data).ConfigureAwait(false)).Result;
 				numberpoolResponse.Interface = this;
 				return numberpoolResponse;
 			});
         }
         public async Task<Numbers> Find_NumberAsync(string uuid,
-			string number)
+			string number, string service = null)
 		{
-        var numberpoolresp = await GetResource<Numbers>("NumberPool/"+uuid+"/Number/"+number);
+        var mandatoryParams = new List<string>{""};
+        var data= CreateData(
+        mandatoryParams,
+        new
+        {
+            service
+        }); 
+        var numberpoolresp = await GetResource<Numbers>("NumberPool/"+uuid+"/Number/"+number, data);
         numberpoolresp.Interface = this;
         return numberpoolresp;