+++ src/Plivo/Resource/Powerpack/Shortcode.cs
@@ -13,7 +13,7 @@ namespace Plivo.Resource.Powerpack {
         public Shortcode () {
 
         }
-        public new string Id => NumberPoolUUID;
+        public new string Id => NumberPoolUuid;
 
         /// <summary>
         /// shortcode
@@ -31,13 +31,13 @@ namespace Plivo.Resource.Powerpack {
         /// Gets or sets the created_on.
         /// </summary>
         /// <value>The created_on.</value>
-        public string CountryISO2 { get; set; }
+        public string CountryIso2 { get; set; }
 
         /// <summary>
         /// Gets or sets the NumberPoolUUID.
         /// </summary>
         /// <value>The NumberPoolUUID.</value>
-        public string NumberPoolUUID { get; set; }
+        public string NumberPoolUuid { get; set; }
 
         // public Shortcode Get(string memberId)
         // {
@@ -82,9 +82,9 @@ namespace Plivo.Resource.Powerpack {
 
         public override string ToString () {
             return "\n" +
-                "NumberPoolUUID: " + NumberPoolUUID + "\n" +
+                "NumberPoolUUID: " + NumberPoolUuid + "\n" +
                 "Shortcode: " + shortcode + "\n" +
-                "CountryISO2: " + CountryISO2 + "\n" +
+                "CountryISO2: " + CountryIso2 + "\n" +
                 "AddedOn: " + AddedOn + "\n";
         }
     }