+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 
 # Change Log
 
+## [v4.14.0](https://github.com/plivo/plivo-dotnet/tree/v4.14.0) (2020-10-27)
+- Change lookup API endpoint and response.
+
+## [v4.13.0](https://github.com/plivo/plivo-dotnet/tree/v4.13.0) (2020-10-06)
+- Add support for Lookup API.
+
+## [v4.12.0](https://github.com/plivo/plivo-dotnet/tree/v4.12.0) (2020-09-24)
+- Add "PublicUri" optional param support for Application API.
+
+## [v4.11.0](https://github.com/plivo/plivo-dotnet/tree/v4.11.0) (2020-09-07)
+- Add Powerpack for MMS
+
 ## [v4.10.2](https://github.com/plivo/plivo-dotnet/tree/v4.10.2) (2020-08-19)
 - Internal changes in PHLO for MultiPartyCall component.